refactor: testimonials between different sections
